requirements.txt
pandas
tqdm
requests
python-dotenv
